mqtt_client/requirements.txt
paho-mqtt==1.6.1
requests==2.25.1
